Added implicit type conversion warnings and "fixed" those warnings by explicit type conversion (in default library and some core files) The reason for adding this compiler flag is to make it more difficult to make type conversion mistakes, even if it will cause unnecessary warnings in many cases

+++ HopsanCLI/ModelUtilities.cpp
@@ -56,7 +56,7 @@ void generateFullSystemHierarchyName(const ComponentSystem *pSys, HString &rFull
     }
 }
 
-HString generateFullPortVariableName(const Port *pPort, const int dataId)
+HString generateFullPortVariableName(const Port *pPort, const size_t dataId)
 {
    HString fullName;
    const NodeDataDescription *pND = pPort->getNodeDataDescription(dataId);
